Improved type resolution in decider registration cargs
Brought in java.beans.Expression class to perform dynamic
constructor resolution. This is in rtesponse to some feedback
from Luc on the SWF team.
