+++ aws-java-sdk-swf-libraries/src/main/java/com/amazonaws/services/simpleworkflow/flow/pojo/POJOWorkflowImplementationFactory.java
@@ -18,10 +18,10 @@ import com.amazonaws.services.simpleworkflow.flow.DecisionContext;
 
 public interface POJOWorkflowImplementationFactory {
     
-    public Object newInstance(DecisionContext decisionContext) throws Exception;
+    Object newInstance(DecisionContext decisionContext) throws Exception;
 
-    public Object newInstance(DecisionContext decisionContext, Object[] constructorArgs) throws Exception;
+    Object newInstance(DecisionContext decisionContext, Object[] constructorArgs) throws Exception;
 
-    public void deleteInstance(Object instance);
+    void deleteInstance(Object instance);
 
 }
